Refactor StartTime and EndTime
